simplified CC workflow
