+++ ilastik/workflows/__init__.py
@@ -66,4 +66,5 @@ if ilastik.config.cfg.getboolean('ilastik', 'debug'):
     import examples.thresholdMasking
     import examples.deviationFromMean
     import examples.labeling
+    import examples.connectedComponents
     import tracking.chaingraph
