+++ ilastik/workflows/examples/connectedComponents/drawer.ui
